+++ lib/Configuration.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * PrivateBin
  *
@@ -7,7 +7,6 @@
  * @link      https://github.com/PrivateBin/PrivateBin
  * @copyright 2012 Sébastien SAUVAGE (sebsauvage.net)
  * @license   https://www.opensource.org/licenses/zlib-license.php The zlib/libpng License
- * @version   1.7.1
  */
 
 namespace PrivateBin;
@@ -40,6 +39,7 @@ class Configuration
             'basepath'                 => '',
             'discussion'               => true,
             'opendiscussion'           => false,
+            'discussiondatedisplay'    => true,
             'password'                 => true,
             'fileupload'               => false,
             'burnafterreadingselected' => false,
@@ -240,7 +240,7 @@ class Configuration
 
         // ensure the basepath ends in a slash, if one is set
         if (
-            strlen($this->_configuration['main']['basepath']) &&
+            !empty($this->_configuration['main']['basepath']) &&
             substr_compare($this->_configuration['main']['basepath'], '/', -1) !== 0
         ) {
             $this->_configuration['main']['basepath'] .= '/';